refactor: make a subcommand
