+++ .changeset/nice-cycles-search.md
@@ -1,6 +1,7 @@
 ---
 '@pnpm/license-scanner': major
 '@pnpm/plugin-commands-licenses': major
+"pnpm": minor
 ---
 
-Added new command named licenses which displays the licenses of the packages
+Added a new command `pnpm licenses list`, which displays the licenses of the packages [#2825](https://github.com/pnpm/pnpm/issues/2825)
